pyproject.toml
[tool.poetry]
name = "genai-crops-analyzer"
version = "0.1.0"
description = "GenAI-powered crops analyzer for detecting nutrient deficiencies and diseases"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
# flask = "^2.3.3"  # Removed - using FastAPI instead
python-dotenv = "^1.0.0"
numpy = ">=1.26.0,<2.0.0"
# pandas = "^2.0.3"  # Removed - not needed for external services
pillow = "^10.0.0"
# scikit-image = "^0.21.0"  # Removed - not needed for external services
# matplotlib = "^3.7.2"  # Removed - not needed for external services  
# scikit-learn = "^1.3.0"  # Removed - not needed for external services
# transformers = "^4.32.1"  # Removed - using external BLIP service
# torch = "^2.0.1"  # Removed - using external services
# torchvision = "^0.15.2"  # Removed - using external services
# opencv-python = "^4.8.0"  # Removed - not needed for external services
# gunicorn = "^21.2.0"  # Removed - using Uvicorn with FastAPI instead
dependency-injector = "^4.46.0"
ollama = "^0.4.8"
# flask-cors = "^6.0.0"  # Removed - using FastAPI CORS middleware instead
sqlalchemy = {extras = ["asyncio"], version = "^2.0.41"}
pydantic = {extras = ["email"], version = "^2.11.5"}
asyncpg = "^0.29.0"
fastapi = "^0.103.2"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
python-multipart = "^0.0.6"
pyyaml = "^6.0.2"
langchain = "^0.1.0"
langchain-core = "^0.1.0"
langchain-community = "^0.0.30"
# asgiref = "^3.7.0"  # Removed - not used in code
email-validator = "^2.2.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
redis = {extras = ["hiredis"], version = "^6.2.0"}
timescale = "^0.0.8"  # Required for TimescaleDB functionality
s3fs = "^2025.5.1"  # Required for weather data from S3
xarray = "^2025.4.0"  # Required for weather data processing
# h5netcdf = "^1.6.1"  # Removed - not used in code
# httpx = "^0.27.0"  # Removed - can use requests instead
aiohttp = "^3.9.0"  # Lighter async HTTP client

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
flake8 = "^6.1.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "src:create_app" 